requirements.txt
# Core Agent Frameworks
langgraph==0.2.34
langgraph-checkpoint>=2.0.0,<3.0.0
crewai==1.9.3
crewai-tools==1.9.3  # ← Changed from 0.23.0 to 1.9.3


# LLM APIs
openai==1.12.0
google-generativeai==0.3.2

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
sse-starlette==1.8.2

# Database
sqlalchemy==2.0.25
aiosqlite==0.20.0

# Visualization
plotly==5.18.0
streamlit==1.31.0
networkx==3.2.1
matplotlib==3.8.2
pandas==2.1.4

# HTTP & APIs
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
rich==13.7.0
loguru==0.7.2

# Data Processing
numpy==1.26.3
